+++ i18n/en.pot
@@ -5,13 +5,52 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2020-05-12T10:00:51.474Z\n"
-"PO-Revision-Date: 2020-05-12T10:00:51.474Z\n"
-
-msgid "{{n}} data values from {{name}} google data set."
-msgstr ""
+"POT-Creation-Date: 2020-05-15T14:48:39.043Z\n"
+"PO-Revision-Date: 2020-05-15T14:48:39.043Z\n"
 
 msgid "No effective import into DHIS2, file download"
+msgstr ""
+
+msgid "Fixed period"
+msgstr ""
+
+msgid "Today"
+msgstr ""
+
+msgid "Yesterday"
+msgstr ""
+
+msgid "Last 7 days"
+msgstr ""
+
+msgid "Last 14 days"
+msgstr ""
+
+msgid "This week"
+msgstr ""
+
+msgid "Last week"
+msgstr ""
+
+msgid "This month"
+msgstr ""
+
+msgid "Last month"
+msgstr ""
+
+msgid "This quarter"
+msgstr ""
+
+msgid "Last quarter"
+msgstr ""
+
+msgid "This year"
+msgstr ""
+
+msgid "Last year"
+msgstr ""
+
+msgid "{{n}} data values from {{name}} google data set."
 msgstr ""
 
 msgid "Import config failed"
@@ -252,45 +291,6 @@ msgid "New mapping"
 msgstr ""
 
 msgid "Edit mapping"
-msgstr ""
-
-msgid "Fixed period"
-msgstr ""
-
-msgid "Today"
-msgstr ""
-
-msgid "Yesterday"
-msgstr ""
-
-msgid "Last 7 days"
-msgstr ""
-
-msgid "Last 14 days"
-msgstr ""
-
-msgid "This week"
-msgstr ""
-
-msgid "Last week"
-msgstr ""
-
-msgid "This month"
-msgstr ""
-
-msgid "Last month"
-msgstr ""
-
-msgid "This quarter"
-msgstr ""
-
-msgid "Last quarter"
-msgstr ""
-
-msgid "This year"
-msgstr ""
-
-msgid "Last year"
 msgstr ""
 
 msgid "Field {{field}} cannot be blank"
