+++ src/webapp/pages/root/Root.tsx
@@ -2,16 +2,15 @@ import React from "react";
 import { Route, Switch, HashRouter } from "react-router-dom";
 import ImportDetail from "../import/ImportDetail";
 import MappingCreationPage from "../mappings/edit-mappings/MappingCreationPage";
+import LandingPage from "../home/HomePage";
 
 const Root = () => {
     return (
         <HashRouter>
             <Switch>
-                <Route
-                    path="/imports/:prefix"
-                    render={({ match }) => <ImportDetail prefix={match.params.prefix} />}
-                />
-                <Route path="/imports" render={() => <ImportDetail prefix="default" />} />
+                <Route path="/" exact render={() => <LandingPage />} />
+
+                <Route path="/import" render={() => <ImportDetail prefix="default" />} />
 
                 <Route path="/mappings/new" render={() => <MappingCreationPage action={"new"} />} />
                 <Route
@@ -20,9 +19,6 @@ const Root = () => {
                         <MappingCreationPage action={"edit"} id={match.params.id} />
                     )}
                 />
-
-                {/* Default route */}
-                <Route render={() => <ImportDetail prefix={"default"} />} />
             </Switch>
         </HashRouter>
     );
